requirements.txt
amqp==2.5.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
decorator==4.4.0
Django==2.2.1
djangorestframework==3.9.4
drf-yasg==1.15.0
factory-boy==2.12.0
Faker==0.9.1
fancycompleter==0.8
httpie==1.0.2
idna==2.8
importlib-metadata==0.17
inflection==0.3.1
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
kombu==4.6.1
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==7.0.0
packaging==19.0
parso==0.4.0
pdbpp==0.10.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.0
pytest==4.6.1
pytest-cov==2.7.1
pytest-django==3.4.8
pytest-factoryboy==2.0.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.22.0
ruamel.yaml==0.15.96
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
wmctrl==0.3
zipp==0.5.1
